requirements.txt
pyyaml
pykeepass>=4.0.6
argcomplete
